requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
emoji==0.5.4
aiogram==2.9.2
joblib==0.14.1
aioredis==1.3.1
torchvision==0.4.2+cpu
requests==2.22.0
selenium==3.141.0
torch==1.3.1+cpu
webdriver_manager==3.2.1
numpy==1.19.1
Pillow==7.2.0
python-dotenv==0.14.0
scipy==1.2.1
